Template refactoring

+++ src/main/kotlin/com/example/captcha/verification/entities/EvaluationResult.kt
@@ -1,3 +1,3 @@
 package com.example.captcha.verification.entities
 
-data class EvaluationResult(val evaluation: Float)
+data class EvaluationResult(val evaluation: Double)
